+++ spec/spec_helper.rb
@@ -1,15 +1,13 @@
 # frozen_string_literal: true
 
 require "agent/lock"
+require "rspec/its"
+
+Dir[File.join(__dir__, "support", "**", "*.rb")].each { |file| require file }
 
 RSpec.configure do |config|
-  # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"
-
-  # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!
-
-  config.expect_with :rspec do |c|
-    c.syntax = :expect
-  end
+  config.expect_with(:rspec) { |c| c.syntax = :expect }
+  config.include_context "a checkout", type: :checkout
 end